projects/stocks-app-py-master/requirements.txt
python-dotenv
ipython
requests
